requirements.txt
langchain
langchain-openai
python-dotenv
langchain_community
langchain-core
ipykernel
streamlit
langchain_groq